requirements.txt
fastapi
uvicorn
openai
pydantic
python-multipart
python-dotenv
httpx
